+++ src/main/java/pl/plajer/buildbattle4/events/GameEvents.java
@@ -64,7 +64,6 @@ import pl.plajer.buildbattle4.arena.ArenaManager;
 import pl.plajer.buildbattle4.arena.ArenaRegistry;
 import pl.plajer.buildbattle4.arena.ArenaState;
 import pl.plajer.buildbattle4.arena.plots.ArenaPlot;
-import pl.plajer.buildbattle4.entities.BuildBattleEntity;
 import pl.plajer.buildbattle4.handlers.ChatManager;
 import pl.plajer.buildbattle4.handlers.items.SpecialItemManager;
 import pl.plajer.buildbattle4.menus.OptionsMenu;
@@ -91,16 +90,32 @@ public class GameEvents implements Listener {
   @EventHandler
   public void onVote(PlayerInteractEvent event) {
     try {
-      if (event.getHand() == EquipmentSlot.OFF_HAND) return;
-      if (event.getAction() == Action.LEFT_CLICK_AIR || event.getAction() == Action.LEFT_CLICK_BLOCK) return;
-      if (event.getItem() == null) return;
+      if (event.getHand() == EquipmentSlot.OFF_HAND) {
+        return;
+      }
+      if (event.getAction() == Action.LEFT_CLICK_AIR || event.getAction() == Action.LEFT_CLICK_BLOCK) {
+        return;
+      }
+      if (event.getItem() == null) {
+        return;
+      }
       Arena arena = ArenaRegistry.getArena(event.getPlayer());
-      if (arena == null) return;
-      if (arena.getArenaState() != ArenaState.IN_GAME) return;
+      if (arena == null) {
+        return;
+      }
+      if (arena.getArenaState() != ArenaState.IN_GAME) {
+        return;
+      }
 
-      if (!event.getItem().hasItemMeta()) return;
-      if (!event.getItem().getItemMeta().hasDisplayName()) return;
-      if (!arena.isVoting()) return;
+      if (!event.getItem().hasItemMeta()) {
+        return;
+      }
+      if (!event.getItem().getItemMeta().hasDisplayName()) {
+        return;
+      }
+      if (!arena.isVoting()) {
+        return;
+      }
       if (arena.getVotingPlot().getOwners().contains(event.getPlayer().getUniqueId())) {
         event.getPlayer().sendMessage(ChatManager.PLUGIN_PREFIX + ChatManager.colorMessage("In-Game.Messages.Voting-Messages.Cant-Vote-Own-Plot"));
         event.setCancelled(true);
@@ -117,16 +132,30 @@ public class GameEvents implements Listener {
   @EventHandler(priority = EventPriority.LOWEST)
   public void onLeave(PlayerInteractEvent event) {
     try {
-      if (event.getHand() == EquipmentSlot.OFF_HAND) return;
-      if (event.getAction() == Action.LEFT_CLICK_AIR || event.getAction() == Action.LEFT_CLICK_BLOCK) return;
+      if (event.getHand() == EquipmentSlot.OFF_HAND) {
+        return;
+      }
+      if (event.getAction() == Action.LEFT_CLICK_AIR || event.getAction() == Action.LEFT_CLICK_BLOCK) {
+        return;
+      }
       Arena arena = ArenaRegistry.getArena(event.getPlayer());
-      if (arena == null) return;
+      if (arena == null) {
+        return;
+      }
       ItemStack itemStack = event.getPlayer().getInventory().getItemInMainHand();
-      if (itemStack == null) return;
-      if (itemStack.getItemMeta() == null) return;
-      if (itemStack.getItemMeta().getDisplayName() == null) return;
+      if (itemStack == null) {
+        return;
+      }
+      if (itemStack.getItemMeta() == null) {
+        return;
+      }
+      if (itemStack.getItemMeta().getDisplayName() == null) {
+        return;
+      }
       String key = SpecialItemManager.getRelatedSpecialItem(itemStack);
-      if (key == null) return;
+      if (key == null) {
+        return;
+      }
       if (SpecialItemManager.getRelatedSpecialItem(itemStack).equalsIgnoreCase("Leave")) {
         event.setCancelled(true);
         if (plugin.isBungeeActivated()) {
@@ -144,17 +173,35 @@ public class GameEvents implements Listener {
   @EventHandler
   public void onOpenOptionMenu(PlayerInteractEvent event) {
     try {
-      if (event.getHand() == EquipmentSlot.OFF_HAND) return;
-      if (event.getAction() == Action.LEFT_CLICK_AIR || event.getAction() == Action.LEFT_CLICK_BLOCK) return;
-      if (event.getItem() == null) return;
+      if (event.getHand() == EquipmentSlot.OFF_HAND) {
+        return;
+      }
+      if (event.getAction() == Action.LEFT_CLICK_AIR || event.getAction() == Action.LEFT_CLICK_BLOCK) {
+        return;
+      }
+      if (event.getItem() == null) {
+        return;
+      }
       Arena arena = ArenaRegistry.getArena(event.getPlayer());
-      if (arena == null) return;
-      if (arena.getArenaState() != ArenaState.IN_GAME) return;
+      if (arena == null) {
+        return;
+      }
+      if (arena.getArenaState() != ArenaState.IN_GAME) {
+        return;
+      }
       ItemStack itemStack = event.getItem();
-      if (!itemStack.hasItemMeta()) return;
-      if (!itemStack.getItemMeta().hasDisplayName()) return;
-      if (arena.isVoting()) return;
-      if (!OptionsMenu.getMenuItem().getItemMeta().getDisplayName().equalsIgnoreCase(itemStack.getItemMeta().getDisplayName())) return;
+      if (!itemStack.hasItemMeta()) {
+        return;
+      }
+      if (!itemStack.getItemMeta().hasDisplayName()) {
+        return;
+      }
+      if (arena.isVoting()) {
+        return;
+      }
+      if (!OptionsMenu.getMenuItem().getItemMeta().getDisplayName().equalsIgnoreCase(itemStack.getItemMeta().getDisplayName())) {
+        return;
+      }
       OptionsMenu.openMenu(event.getPlayer(), arena.getPlotManager().getPlot(event.getPlayer()));
     } catch (Exception ex) {
       new ReportedException(plugin, ex);
@@ -180,9 +227,13 @@ public class GameEvents implements Listener {
 
   @EventHandler
   public void onFoodChange(FoodLevelChangeEvent event) {
-    if (!(event.getEntity().getType() == EntityType.PLAYER)) return;
+    if (!(event.getEntity().getType() == EntityType.PLAYER)) {
+      return;
+    }
     Player player = (Player) event.getEntity();
-    if (ArenaRegistry.getArena(player) == null) return;
+    if (ArenaRegistry.getArena(player) == null) {
+      return;
+    }
     event.setCancelled(true);
     player.setFoodLevel(20);
   }
@@ -225,15 +276,23 @@ public class GameEvents implements Listener {
   @EventHandler
   public void onTNTInteract(PlayerInteractEvent event) {
     try {
-      if (event.getHand() == EquipmentSlot.OFF_HAND) return;
+      if (event.getHand() == EquipmentSlot.OFF_HAND) {
+        return;
+      }
       Player player = event.getPlayer();
       Arena arena = ArenaRegistry.getArena(player);
-      if (arena == null) return;
-      if (player.getInventory().getItemInMainHand() == null) return;
+      if (arena == null) {
+        return;
+      }
+      if (player.getInventory().getItemInMainHand() == null) {
+        return;
+      }
       if (player.getInventory().getItemInMainHand().getType() != Material.FLINT_AND_STEEL) {
         return;
       }
-      if (event.getClickedBlock() == null) return;
+      if (event.getClickedBlock() == null) {
+        return;
+      }
       if (event.getClickedBlock().getType() == Material.TNT) {
         event.setCancelled(true);
       }
@@ -245,10 +304,14 @@ public class GameEvents implements Listener {
   @EventHandler
   public void onEntityDamageEntity(EntityDamageByEntityEvent event) {
     try {
-      if (event.getEntity().getType() != EntityType.PLAYER) return;
+      if (event.getEntity().getType() != EntityType.PLAYER) {
+        return;
+      }
       Player player = (Player) event.getEntity();
       Arena arena = ArenaRegistry.getArena(player);
-      if (arena == null) return;
+      if (arena == null) {
+        return;
+      }
       event.setCancelled(true);
     } catch (Exception ex) {
       new ReportedException(plugin, ex);
@@ -259,11 +322,17 @@ public class GameEvents implements Listener {
   public void onTreeGrow(StructureGrowEvent event) {
     try {
       Arena arena = ArenaRegistry.getArena(event.getPlayer());
-      if (arena == null) return;
+      if (arena == null) {
+        return;
+      }
       ArenaPlot buildPlot = arena.getPlotManager().getPlot(event.getPlayer());
-      if (buildPlot == null) return;
+      if (buildPlot == null) {
+        return;
+      }
       for (BlockState blockState : event.getBlocks()) {
-        if (!buildPlot.getCuboid().isIn(blockState.getLocation())) blockState.setType(Material.AIR);
+        if (!buildPlot.getCuboid().isIn(blockState.getLocation())) {
+          blockState.setType(Material.AIR);
+        }
       }
     } catch (Exception ex) {
       new ReportedException(plugin, ex);
@@ -318,24 +387,36 @@ public class GameEvents implements Listener {
   public void onOptionMenuClick(InventoryClickEvent e) {
     try {
       if (e.getWhoClicked() instanceof Player && ArenaRegistry.getArena((Player) e.getWhoClicked()) != null && e.getCurrentItem() != null &&
-              e.getCurrentItem().getType() == Material.NETHER_STAR && e.getCurrentItem().getItemMeta().hasDisplayName() &&
-              e.getCurrentItem().getItemMeta().getDisplayName().equals(ChatManager.colorMessage("Menus.Option-Menu.Option-Item"))) {
+          e.getCurrentItem().getType() == Material.NETHER_STAR && e.getCurrentItem().getItemMeta().hasDisplayName() &&
+          e.getCurrentItem().getItemMeta().getDisplayName().equals(ChatManager.colorMessage("Menus.Option-Menu.Option-Item"))) {
         e.setResult(Event.Result.DENY);
         e.setCancelled(true);
       }
-      if (e.getInventory() == null) return;
-      if (e.getCurrentItem() == null) return;
-      if (!e.getCurrentItem().hasItemMeta()) return;
-      if (!e.getCurrentItem().getItemMeta().hasDisplayName()) return;
+      if (e.getInventory() == null) {
+        return;
+      }
+      if (e.getCurrentItem() == null) {
+        return;
+      }
+      if (!e.getCurrentItem().hasItemMeta()) {
+        return;
+      }
+      if (!e.getCurrentItem().getItemMeta().hasDisplayName()) {
+        return;
+      }
       String displayName = e.getCurrentItem().getItemMeta().getDisplayName();
       Player player = (Player) e.getWhoClicked();
       if (e.getInventory().getName().equals(ChatManager.colorMessage("Menus.Option-Menu.Inventory-Name"))) {
         e.setCancelled(true);
       }
       Arena arena = ArenaRegistry.getArena((Player) e.getWhoClicked());
-      if (arena == null) return;
+      if (arena == null) {
+        return;
+      }
 
-      if (arena.getArenaState() != ArenaState.IN_GAME) return;
+      if (arena.getArenaState() != ArenaState.IN_GAME) {
+        return;
+      }
       if (displayName.equalsIgnoreCase(ChatManager.colorMessage("Menus.Option-Menu.Particle-Option"))) {
         e.getWhoClicked().closeInventory();
         ParticleMenu.openMenu(player);
@@ -379,7 +460,7 @@ public class GameEvents implements Listener {
         return;
       }
       if (e.getCursor().getType() == null || e.getCursor().getType() == Material.SAPLING || e.getCursor().getType() == Material.TRAP_DOOR || e.getCursor().getType() == Material.WOOD_DOOR || e.getCursor().getType() == Material.IRON_TRAPDOOR || e.getCursor().getType() == Material.WOODEN_DOOR || e.getCursor().getType() == Material.ACACIA_DOOR || e.getCursor().getType() == Material.BIRCH_DOOR || e.getCursor().getType() == Material.WOOD_DOOR || e.getCursor().getType() == Material.JUNGLE_DOOR || e.getCursor().getType() == Material.SPRUCE_DOOR || e.getCursor().getType() == Material.IRON_DOOR || e.getCursor().getType() == Material.CHEST || e.getCursor().getType() == Material.TRAPPED_CHEST || e.getCursor().getType() == Material.FENCE_GATE || e.getCursor().getType() == Material.BED || e.getCursor().getType() == Material.LADDER || e.getCursor().getType() == Material.JUNGLE_FENCE_GATE || e.getCursor().getType() == Material.JUNGLE_DOOR_ITEM || e.getCursor().getType() == Material.SIGN || e.getCursor().getType() == Material.SIGN_POST || e.getCursor().getType() == Material.WALL_SIGN || e.getCursor().getType() == Material.CACTUS || e.getCursor().getType() == Material.ENDER_CHEST || e.getCursor().getType() == Material.PISTON_BASE
-              || e.getCursor().getType() == Material.TNT || e.getCursor().getType() == Material.AIR) {
+          || e.getCursor().getType() == Material.TNT || e.getCursor().getType() == Material.AIR) {
         e.setCancelled(true);
         return;
       }
@@ -404,12 +485,20 @@ public class GameEvents implements Listener {
   @EventHandler(priority = EventPriority.HIGHEST)
   public void onPreCommand(PlayerCommandPreprocessEvent event) {
     try {
-      if (ArenaRegistry.getArena(event.getPlayer()) == null) return;
-      for (String string : ConfigPreferences.getWhitelistedCommands()) {
-        if (event.getMessage().contains(string)) return;
+      if (ArenaRegistry.getArena(event.getPlayer()) == null) {
+        return;
       }
-      if (event.getPlayer().isOp() || event.getPlayer().hasPermission("buildbattle.admin") || event.getPlayer().hasPermission("buildbattle.command.bypass")) return;
-      if (event.getMessage().contains("leave") || event.getMessage().contains("stats")) return;
+      for (String string : ConfigPreferences.getWhitelistedCommands()) {
+        if (event.getMessage().contains(string)) {
+          return;
+        }
+      }
+      if (event.getPlayer().isOp() || event.getPlayer().hasPermission("buildbattle.admin") || event.getPlayer().hasPermission("buildbattle.command.bypass")) {
+        return;
+      }
+      if (event.getMessage().contains("leave") || event.getMessage().contains("stats")) {
+        return;
+      }
       event.setCancelled(true);
       event.getPlayer().sendMessage(ChatManager.colorMessage("In-Game.Only-Command-Ingame-Is-Leave"));
     } catch (Exception ex) {
@@ -421,10 +510,16 @@ public class GameEvents implements Listener {
   public void onBucketEmpty(PlayerBucketEmptyEvent event) {
     try {
       Arena arena = ArenaRegistry.getArena(event.getPlayer());
-      if (arena == null) return;
+      if (arena == null) {
+        return;
+      }
       ArenaPlot buildPlot = arena.getPlotManager().getPlot(event.getPlayer());
-      if (buildPlot == null) return;
-      if (!buildPlot.getCuboid().isIn(event.getBlockClicked().getRelative(event.getBlockFace()).getLocation())) event.setCancelled(true);
+      if (buildPlot == null) {
+        return;
+      }
+      if (!buildPlot.getCuboid().isIn(event.getBlockClicked().getRelative(event.getBlockFace()).getLocation())) {
+        event.setCancelled(true);
+      }
     } catch (Exception ex) {
       new ReportedException(plugin, ex);
     }
@@ -436,7 +531,9 @@ public class GameEvents implements Listener {
       for (Arena arena : ArenaRegistry.getArenas()) {
         if (arena.getPlotManager().getPlots().size() != 0 && arena.getPlotManager().getPlots().get(0) != null) {
           if (arena.getPlotManager().getPlots().get(0).getCuboid().getCenter().getWorld().getName().equalsIgnoreCase(event.getBlock().getWorld().getName())) {
-            if (event.getSource().getType() == Material.FIRE) event.setCancelled(true);
+            if (event.getSource().getType() == Material.FIRE) {
+              event.setCancelled(true);
+            }
           }
         }
       }
@@ -450,7 +547,9 @@ public class GameEvents implements Listener {
     try {
       for (Arena arena : ArenaRegistry.getArenas()) {
         if (arena.getStartLocation() != null && event.getEntity().getWorld().equals(arena.getStartLocation().getWorld())) {
-          if (event.getSpawnReason() == CreatureSpawnEvent.SpawnReason.CUSTOM) return;
+          if (event.getSpawnReason() == CreatureSpawnEvent.SpawnReason.CUSTOM) {
+            return;
+          }
           if (event.getEntity().getType() == EntityType.WITHER || ConfigPreferences.isMobSpawningDisabled()) {
             event.setCancelled(true);
             return;
@@ -471,7 +570,7 @@ public class GameEvents implements Listener {
               } else {
                 buildplot.addEntity();
                 event.setCancelled(false);
-                new BuildBattleEntity(event.getEntity()).toggleMoveable();
+                event.getEntity().setAI(false);
               }
             }
           }
@@ -497,13 +596,22 @@ public class GameEvents implements Listener {
   @EventHandler
   public void onPlayerDropItem(PlayerDropItemEvent event) {
     try {
-      if (ArenaRegistry.getArena(event.getPlayer()) == null) return;
-      if (event.getItemDrop().getItemStack() == null) return;
+      if (ArenaRegistry.getArena(event.getPlayer()) == null) {
+        return;
+      }
+      if (event.getItemDrop().getItemStack() == null) {
+        return;
+      }
       ItemStack drop = event.getItemDrop().getItemStack();
-      if (!drop.hasItemMeta()) return;
-      if (!drop.getItemMeta().hasDisplayName()) return;
-      if (drop.getItemMeta().getDisplayName().equals(ChatManager.colorMessage("Menus.Option-Menu.Inventory-Name")) || VoteItems.getPoints(drop) != 0)
+      if (!drop.hasItemMeta()) {
+        return;
+      }
+      if (!drop.getItemMeta().hasDisplayName()) {
+        return;
+      }
+      if (drop.getItemMeta().getDisplayName().equals(ChatManager.colorMessage("Menus.Option-Menu.Inventory-Name")) || VoteItems.getPoints(drop) != 0) {
         event.setCancelled(true);
+      }
     } catch (Exception ex) {
       new ReportedException(plugin, ex);
     }
@@ -514,7 +622,9 @@ public class GameEvents implements Listener {
   public void onBreak(BlockBreakEvent event) {
     try {
       Arena arena = ArenaRegistry.getArena(event.getPlayer());
-      if (arena == null) return;
+      if (arena == null) {
+        return;
+      }
       if (arena.getArenaState() != ArenaState.IN_GAME) {
         event.setCancelled(true);
         return;
@@ -547,7 +657,9 @@ public class GameEvents implements Listener {
   public void onPlace(BlockPlaceEvent event) {
     try {
       Arena arena = ArenaRegistry.getArena(event.getPlayer());
-      if (arena == null) return;
+      if (arena == null) {
+        return;
+      }
       if (arena.getArenaState() != ArenaState.IN_GAME) {
         event.setCancelled(true);
         return;
@@ -580,7 +692,9 @@ public class GameEvents implements Listener {
   public void onInventoryClick(InventoryClickEvent event) {
     try {
       Arena arena = ArenaRegistry.getArena((Player) event.getWhoClicked());
-      if (arena == null) return;
+      if (arena == null) {
+        return;
+      }
       if (arena.getArenaState() != ArenaState.IN_GAME) {
         event.setCancelled(true);
         return;
@@ -598,7 +712,9 @@ public class GameEvents implements Listener {
   public void onChat(AsyncPlayerChatEvent event) {
     if (ArenaRegistry.getArena(event.getPlayer()) == null) {
       for (Player player : event.getRecipients()) {
-        if (ArenaRegistry.getArena(event.getPlayer()) == null) return;
+        if (ArenaRegistry.getArena(event.getPlayer()) == null) {
+          return;
+        }
         event.getRecipients().remove(player);
       }
     }
@@ -609,15 +725,29 @@ public class GameEvents implements Listener {
   @EventHandler
   public void onNPCClick(PlayerInteractEntityEvent e) {
     try {
-      if (e.getHand() == EquipmentSlot.OFF_HAND) return;
-      if (e.getPlayer().getInventory().getItemInMainHand() == null || e.getPlayer().getInventory().getItemInMainHand().getType() == Material.AIR) return;
+      if (e.getHand() == EquipmentSlot.OFF_HAND) {
+        return;
+      }
+      if (e.getPlayer().getInventory().getItemInMainHand() == null || e.getPlayer().getInventory().getItemInMainHand().getType() == Material.AIR) {
+        return;
+      }
       if (e.getRightClicked() instanceof Villager && e.getRightClicked().getCustomName() != null && e.getRightClicked().getCustomName().equalsIgnoreCase(ChatManager.colorMessage("In-Game.NPC.Floor-Change-NPC-Name"))) {
         Arena arena = ArenaRegistry.getArena(e.getPlayer());
-        if (arena == null) return;
-        if (!e.getPlayer().getInventory().getItemInMainHand().getType().isBlock()) return;
-        if (arena.getBlacklistedBlocks().contains(e.getPlayer().getInventory().getItemInMainHand().getType())) return;
-        if (arena.getArenaState() != ArenaState.IN_GAME) return;
-        if (arena.isVoting()) return;
+        if (arena == null) {
+          return;
+        }
+        if (!e.getPlayer().getInventory().getItemInMainHand().getType().isBlock()) {
+          return;
+        }
+        if (arena.getBlacklistedBlocks().contains(e.getPlayer().getInventory().getItemInMainHand().getType())) {
+          return;
+        }
+        if (arena.getArenaState() != ArenaState.IN_GAME) {
+          return;
+        }
+        if (arena.isVoting()) {
+          return;
+        }
         arena.getPlotManager().getPlot(e.getPlayer()).changeFloor(e.getPlayer().getInventory().getItemInMainHand().getType(), e.getPlayer().getInventory().getItemInMainHand().getData().getData());
         e.getPlayer().sendMessage(ChatManager.PLUGIN_PREFIX + ChatManager.colorMessage("Menus.Option-Menu.Floor-Changed"));
       }